backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Polyvox - Voice accent training"
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "fastapi>=0.129.2",
    "uvicorn[standard]>=0.41.0",
    "boto3>=1.35.0",
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
    "librosa>=0.10.0",
    "pydub>=0.25.0",
    "faster-whisper>=0.9.0",
    "whisper-timestamped>=1.12.0",
    "inflect>=7.0.0",
    "unidecode>=1.3.0",
    "pypinyin>=0.50.0",
    "cn2an>=0.5.0",
    "jieba>=0.42.0",
    "eng-to-ipa>=0.0.2",
    "wavmark>=0.0.3",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.9",
    "huggingface-hub>=0.20.0",
    "onnxruntime>=1.16.0,<1.20.0",
]

[tool.setuptools]
py-modules = []
packages = []